requirements_01.txt
-f https://download.pytorch.org/whl/cu113/torch_stable.html
-f https://data.dgl.ai/wheels/repo.html
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
chainer==7.8.1
chainer-chemistry==0.7.1
charset-normalizer==2.0.11
contextvars==2.4
cssselect2==0.4.1
cycler==0.11.0
decorator==4.4.2
defusedxml==0.7.1
dgl-cu113==0.8.0
dglgo==0.0.1
einops==0.4.0
entrypoints==0.4
filelock==3.4.1
googledrivedownloader==0.4
h5py==3.1.0
idna==3.3
imageio==2.15.0
immutables==0.16
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.5.1
notebook==6.4.8
numpy==1.19.5
opencv-python==4.5.5.62
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkg_resources==0.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.0
pysmiles==1.0.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
rdflib==5.0.0
rdkit-pypi==2021.9.4
requests==2.27.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tinycss2==1.1.1
torch==1.10.2+cu113
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing==3.7.4.3
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
yacs==0.1.8
zipp==3.6.0

